move syncer methods to separate file
